Implemented ticket #543 "Improve error messages"
git-svn-id: https://svn.code.sf.net/p/numere/source/trunk@798 d50cd26f-24d5-421e-9e7d-30b8b4216d91

+++ kernel/core/utils/tools.cpp
@@ -1911,7 +1911,7 @@ long long int intCast(double number)
     if (fabs(number - rint(number)) < 1e-7)
         return rint(number);
     // otherwise truncate
-    return (long long int)number;
+    return static_cast<int>(number);
 }
 
 // This function is a wrapper for the usual wcstombs function, which can handle wstrings
